# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Mercadona IT API
  version: 1.0.0
  description: Aplicación para la gestión de tiendas, secciones, trabajadores y sus asignaciones.
paths:
  /tienda:
    post:
      operationId: insertTienda
      summary: Crea una tienda
      description: Crea una tienda y la asocia automáticamente con todas las secciones existentes
      tags:
        - Tienda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TiendaDto'
      responses:
        '201':
          description: Tienda creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TiendaDto'
        '400':
          description: Datos de la tienda no válidos
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Ya existe una tienda con el código indicado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /tienda/{tiendaId}:
    parameters:
      - name: tiendaId
        in: path
        required: true
        description: Código de la tienda
        schema:
          type: integer
          format: int64
          minimum: 1
        example: 1
    delete:
      operationId: deleteTienda
      summary: Elimina una tienda
      description: Elimina la tienda y sus asociaciones con secciones, siempre que no tenga trabajadores
      tags:
        - Tienda
      responses:
        '204':
          description: Tienda eliminada
        '404':
          description: Tienda no encontrada
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: La tienda no se puede eliminar porque tiene trabajadores
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  schemas:
    ProblemDetail:
      type: object
      required:
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Resource not found
        status:
          type: integer
          format: int32
          example: 404
        detail:
          type: string
          example: Trabajador with identifier 11111111A was not found
        instance:
          type: string
          format: uri
          example: /trabajador/11111111A
...
    TiendaDto:
      type: object
      additionalProperties: false
      description: Tienda de la cadena
      required:
        - codigo
        - nombre
      properties:
        codigo:
          type: integer
          format: int64
          minimum: 1
          description: Código de la tienda
          example: 4
        nombre:
          type: string
          minLength: 1
          maxLength: 100
          description: Nombre de la tienda
          example: Tienda 4
...